refactor: 인프라 관련 커스텀 예외 클래스 패키지 수정
