+++ backend/src/main/java/com/woowacourse/kkogkkog/exception/infrastructure/PostMessageRequestFailedException.java
@@ -1,8 +1,8 @@
-package com.woowacourse.kkogkkog.exception.coupon;
+package com.woowacourse.kkogkkog.exception.infrastructure;
 
 public class PostMessageRequestFailedException extends RuntimeException {
 
     public PostMessageRequestFailedException() {
-        super("슬랙 DM 알림 요청을 실패하였습니다.");
+        super("슬랙 DM 전송 요청에 실패하였습니다.");
     }
 }
