+++ backend/src/main/java/com/woowacourse/kkogkkog/infrastructure/SlackClient.java
@@ -1,6 +1,6 @@
 package com.woowacourse.kkogkkog.infrastructure;
 
-import com.woowacourse.kkogkkog.exception.coupon.PostMessageRequestFailedException;
+import com.woowacourse.kkogkkog.exception.infrastructure.PostMessageRequestFailedException;
 import com.woowacourse.kkogkkog.exception.auth.AccessTokenRequestFailedException;
 import com.woowacourse.kkogkkog.exception.auth.AccessTokenRetrievalFailedException;
 import com.woowacourse.kkogkkog.exception.auth.BotInstallationFailedException;
